issue #484: implement SegmentMerger SPI + aggressive policy + event-driven scheduling
The optimizer pod was running with `NoopMerger` so every merge cycle ended in
`mergeDeclinedTotal++`. On the gist1m k3s-local benchmark this left the index
at 8 ACTIVE segments forever. This change ships a production graph-merge
implementation, makes the merge policy aggressive enough to actually converge
the layout, and replaces the time-based tick with an event-driven path so
sustained ingest doesn't starve the optimizer.

Key changes:

* `RemoteSegmentGraphMerger` (herddb-core): standalone graph-merge utility
  that reads each input's map file from remote storage, drops tombstoned
  ordinals via `TombstoneOverlayManager.loadOverlay`, de-duplicates
  surviving entries by PK preferring the highest-generation source, builds
  a fresh `GraphIndexBuilder` over the survivors, writes a FusedPQ +
  InlineVectors `OnDiskGraphIndex` plus a companion map file, and uploads
  both via the multipart API. Round-trip recall@10 >= 0.9 on a 1500-vector
  EUCLIDEAN dataset.

* `RemoteSegmentMerger` (herddb-indexing-service): `SegmentMerger` impl that
  wraps `RemoteSegmentGraphMerger`, pre-loads each input's overlay through
  the IS-side `TombstoneOverlayManager`, and produces a fresh ACTIVE
  `SegmentMetadata` (random 63-bit `segmentId`, generation = `max(input)+1`).
  Refuses to merge if an overlay can't be loaded — never resurrect
  tombstoned vectors silently. `IndexOptimizerMain.start()` constructs it
  by default when remote-file servers are configured (static or via ZK
  discovery); falls back to `NoopMerger` only when no servers are reachable.

* `MergePolicy.AggressivePolicy`: new default policy. Filters active
  segments to those below `targetMaxBytes` (default 8 GiB — graduated
  segments are left alone), fires whenever 2+ sub-target candidates exist,
  and caps the picked set at `perCycleMaxBytes` and `maxCount`. There is
  no minimum-bytes / minimum-count gating — the optimizer keeps merging
  until every graph has reached the cap. `SmallestFirstPolicy` retained
  verbatim for backward compatibility with existing tests.

* Event-driven scheduling: `IndexOptimizerMain` arms a single
  `ZooKeeper.addWatch(tablespacePath, …, AddWatchMode.PERSISTENT_RECURSIVE)`
  on startup. Every children-changed event under the registry tablespace
  path schedules a debounced (default 500 ms) tick on the engine's
  single-threaded scheduler; bursts of N events coalesce into one tick via
  an `AtomicBoolean` CAS. The existing periodic `scheduleAtFixedRate(…,
  intervalMs)` continues as a safety net. Persistent-recursive watches
  fire forever without re-arming, so a single setup call covers the
  pod's entire lifetime.

* `SegmentMetadata.segmentId` widened from `int` to `long` (issue #484
  user-requested) so the merger can mint full 63-bit random ids without
  collision-avoidance offsets. JSON wire format is forward-compatible.
  New `overlayGeneration` field added so the optimizer reaper and merger
  can reconstruct the exact `tombstones-{N}` multipart file rather than
  brute-forcing the generation window.

* Helm: `targetMaxBytes`, `eventDebounceMs`, `merger.{dim,M,beamWidth,…}`,
  and `remoteFileClient.{servers,timeout,retries,…}` plumbed through
  `values.yaml` and the optimizer configmap; the k3s-local example sets
  `merger.dim=960` for GIST1M.

Tests added:

* `RemoteSegmentGraphMergerTest` — round-trip on 1500 vectors (3 segments,
  dim=64, EUCLIDEAN). Asserts every PK survives, recall@10 >= 0.9 vs a
  brute-force baseline, all-tombstoned input → null output, and
  `deleteOutput()` removes the uploaded multipart pair.
* `RemoteSegmentGraphMergerTombstoneFilterTest` — half-tombstoned input A
  drops its first 50 ordinals; merged segment carries exactly the 50
  surviving A PKs plus all 100 B PKs. Also covers the "every ordinal in A
  tombstoned" case.
* `RemoteSegmentMergerTest` — end-to-end merger SPI: ACTIVE output, random
  segmentId, sizeBytes, vectorCount, tombstone fields cleared, files
  addressable on the DSM. `abandon()` removes both files. Cross-tablespace
  / no-segmentId inputs rejected.
* `AggressivePolicyTest` — fires with two sub-target segments; excludes
  graduated segments; respects `perCycleMaxBytes` and `maxCount` caps;
  always picks at least 2 to keep merging unstuck; smallest-first ordering;
  arg validation.
* `IndexOptimizerEventDrivenTest` — boots `IndexOptimizerMain` against a
  `TestingServer` with `intervalMs=1h` (periodic tick effectively
  disabled), publishes 4 segments, asserts an event-driven tick fires
  within 5 s. Second burst proves the persistent-recursive watch
  survives without re-arming. Debounce coalescing test publishes 10
  segments inside a single window and asserts the tick count is far
  smaller than the event count.
* `IndexOptimizerMainConfigTest` — preconfigured merger used verbatim;
  no-servers fallback to `NoopMerger`; legacy `loadMergerSpi()` helper
  preserved.
* `IndexOptimizerKubernetesIT` updated to set `merger.dim=64` so the
  production merger wiring is exercised end-to-end (the IT does not
  ingest data so a "segment count decreases" assertion is left to the
  bench agents).

Validation:

* Pre-PR: `mvn -B checkstyle:check apache-rat:check spotbugs:check
  install -DskipTests -Pci` clean.
* Hammer: `DirectMultipleConcurrentUpdatesSuite{NoIndexes,WithNonUniqueIndexes,
  WithUniqueIndexes}Test` and `BLinkConcurrentSearchInsertTest` green twice.
* Optimizer suite: 106 tests pass (existing 86 + 20 new).

+++ herddb-indexing-service/src/test/java/herddb/indexing/optimizer/UnsafeModeReaperDeletesMergerOutputTest.java
@@ -108,7 +108,7 @@ public class UnsafeModeReaperDeletesMergerOutputTest {
         }
     }
 
-    private SegmentMetadata sampleSegment(String uuid, int segmentId) {
+    private SegmentMetadata sampleSegment(String uuid, long segmentId) {
         return SegmentMetadata.builder()
                 .segmentUuid(uuid).tablespaceUuid(TS_UUID).tableName("docs")
                 .indexUuid(IDX_UUID).indexName("docs_v1").state(SegmentState.ACTIVE)
@@ -138,7 +138,7 @@ public class UnsafeModeReaperDeletesMergerOutputTest {
         public SegmentMetadata merge(List<SegmentMetadata> inputs, int newOwnerInstance)
                 throws Exception {
             String mergedUuid = UUID.randomUUID().toString();
-            int mergedId = (int) nextSegmentId.getAndIncrement();
+            long mergedId = nextSegmentId.getAndIncrement();
             String multipartUuid = inputs.get(0).getIndexUuid() + "_seg" + mergedId;
 
             Path graph = Files.createTempFile(tmpDir, "graph-", ".bin");
@@ -179,7 +179,7 @@ public class UnsafeModeReaperDeletesMergerOutputTest {
         // (under the indexUUID + "_seg" + segmentId multipart path the reaper will
         // probe). This makes the "files deleted at reap" assertion below meaningful.
         for (int i = 0; i < 3; i++) {
-            int segId = 100 + i;
+            long segId = 100L + i;
             registry.createSegment(sampleSegment("input-" + i, segId));
             String multipartUuid = IDX_UUID + "_seg" + segId;
             Path graph = Files.createTempFile(tmpDir, "input-graph-", ".bin");
@@ -213,7 +213,7 @@ public class UnsafeModeReaperDeletesMergerOutputTest {
         VersionedSegmentMetadata mergedOutput = all.stream()
                 .filter(v -> v.metadata().getState() == SegmentState.ACTIVE)
                 .findFirst().orElseThrow();
-        int mergedSegmentId = mergedOutput.metadata().getSegmentId();
+        long mergedSegmentId = mergedOutput.metadata().getSegmentId();
         assertTrue("merger must produce a real segmentId, not the sentinel",
                 mergedSegmentId != SegmentMetadata.NO_SEGMENT_ID);
